# perf: cache the lookup of should_trace_cache