--- a/coverage/ctracer/stats.h
+++ b/coverage/ctracer/stats.h
@@ -22,6 +22,7 @@ typedef struct Stats {
     unsigned int stack_reallocs;
     unsigned int errors;
     unsigned int pycalls;
+    unsigned int filename_cache_hits;
     unsigned int start_context_calls;
 #endif
 } Stats;