@nedbat 2022-06-11
This is about what I expected, since the dict lookup is only once per function call, and dict lookups are fast.